novos dados primeiro check

+++ src/app/models/DepositMoney.js
@@ -9,7 +9,7 @@ class DepositMoney extends Model {
         type_transaction: Sequelize.STRING,
         local: Sequelize.STRING,
         type_bank: Sequelize.STRING,
-        value: Sequelize.INTEGER,
+        value: Sequelize.BIGINT,
         proof_img: Sequelize.STRING,
       },
       {
